Refactor rule tests into specific and generic tests
